pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minimal-agent-framework"
version = "0.1.0"
description = "A minimal framework for building agent-based applications."
readme = "README.md"
authors = [{ name = "Randy Herritt", email = "randy.herritt@gmail.com" }]
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["agent", "framework", "multi-agent", "ai", "artificial intelligence"]

dependencies = [
  "openai>=1.60.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
]

[project.urls]
Repository = "https://github.com/thepeacefulprogrammer/agent_framework"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["minimal_agent_framework*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
]
